feat(vscode): Add shortcut key for cancelling cell execution and Update workbench settings
Add a new keybinding to cancel the execution of a cell in a notebook when the 's' key is pressed. This provides a quicker way to stop a cell from running without interrupting the flow of work.

Changed the location of the activity bar to 'hidden' for a less cluttered workspace. Also updated the color theme to 'Default Light Modern' for a brighter and more modern look.